+++ src/intelli/gui/graphscene.cpp
@@ -11,17 +11,19 @@
 #include "intelli/gui/graphscene.h"
 
 #include <intelli/graph.h>
-#include "intelli/connection.h"
-#include "intelli/nodefactory.h"
-#include "intelli/nodedatafactory.h"
-#include "intelli/node/groupinputprovider.h"
-#include "intelli/node/groupoutputprovider.h"
-#include "intelli/gui/nodeui.h"
-#include "intelli/gui/nodegeometry.h"
+#include <intelli/connection.h>
+#include <intelli/nodefactory.h>
+#include <intelli/nodedatafactory.h>
+#include <intelli/node/groupinputprovider.h>
+#include <intelli/node/groupoutputprovider.h>
+#include <intelli/gui/nodeui.h>
+#include <intelli/gui/nodegeometry.h>
 #include <intelli/gui/graphscenedata.h>
 #include <intelli/gui/graphics/nodeobject.h>
 #include <intelli/gui/graphics/connectionobject.h>
+#include <intelli/gui/graphics/popupitem.h>
 #include <intelli/private/utils.h>
+#include <intelli/private/gui_utils.h>
 
 #include <gt_application.h>
 #include <gt_command.h>
@@ -81,6 +83,32 @@ enum SelectionFilter
     NodesOnly,
     ConnectionsOnly
 };
+
+/**
+ * @brief Helper method that checks if a node is user deletable
+ * @param o Object
+ * @return Whether the object is user deletable
+ */
+static bool
+isNotDeletable(NodeGraphicsObject* o)
+{
+    return !(o->node().objectFlags() & GtObject::ObjectFlag::UserDeletable);
+};
+
+/**
+ * @brief Returns the currently active view
+ * @param scene Graph Scene
+ * @return Active view (may be null)
+ */
+static QGraphicsView const*
+activeView(GraphScene const* scene)
+{
+    auto const& views = scene->views();
+    auto iter = std::find_if(views.begin(), views.end(), [](QGraphicsView const* view){
+        return view->hasFocus();
+    });
+    return (iter != views.end()) ? *iter : nullptr;
+}
 
 /**
  * @brief Helper function to find all selected nodes and connections. If a
@@ -369,9 +397,32 @@ copySelectionTo(GraphScene& scene, Graph& dummy)
     auto selected = Impl::findSelectedItems(scene);
     if (selected.nodes.empty()) return false;
 
+    // remove unqiue nodes
+    auto const isUnique = [](NodeGraphicsObject const* o){
+        return o->node().nodeFlags() & NodeFlag::Unique;
+    };
+
+    selected.nodes.erase(
+        std::remove_if(selected.nodes.begin(),
+                       selected.nodes.end(),
+                       isUnique),
+        selected.nodes.end());
+
+    // remove not deletable nodes
+    selected.nodes.erase(
+        std::remove_if(selected.nodes.begin(),
+                       selected.nodes.end(),
+                       isNotDeletable),
+        selected.nodes.end());
+
+    // at least one node should be selected
+    if (selected.nodes.empty()) return false;
+
     // only duplicate internal connections
     auto const containsNode = [&selected](NodeId nodeId){
-        return std::find_if(selected.nodes.begin(), selected.nodes.end(), [nodeId](NodeGraphicsObject* o){
+        return std::find_if(selected.nodes.begin(),
+                            selected.nodes.end(),
+                            [nodeId](NodeGraphicsObject* o){
                    return o->nodeId() == nodeId;
                }) != selected.nodes.end();
     };
@@ -386,20 +437,6 @@ copySelectionTo(GraphScene& scene, Graph& dummy)
                        selected.connections.end(),
                        isExternalConnection),
         selected.connections.end());
-
-    // remove unqiue nodes
-    auto const isUnique = [](NodeGraphicsObject const* o){
-        return o->node().nodeFlags() & NodeFlag::Unique;
-    };
-
-    selected.nodes.erase(
-        std::remove_if(selected.nodes.begin(),
-                       selected.nodes.end(),
-                       isUnique),
-        selected.nodes.end());
-
-    // at least one node should be selected
-    if (selected.nodes.empty()) return false;
 
     // append nodes and connections to dummy graph
     bool success = true;
@@ -739,12 +776,70 @@ GraphScene::alignObjectsToGrid()
 void
 GraphScene::deleteSelectedObjects()
 {
-    auto const& selected = Impl::findSelectedItems(*this);
+    auto selected = Impl::findSelectedItems(*this);
     if (selected.empty()) return;
 
+    // remove nodes that are not deletable
+    auto beginErase = std::remove_if(selected.nodes.begin(),
+                                     selected.nodes.end(),
+                                     Impl::isNotDeletable);
+
+    // not all nodes are default deletable
+    int count = std::distance(beginErase, selected.nodes.end());
+    if (count == selected.nodes.size() && !selected.nodes.empty())
+    {
+        if (count == 1)
+        {
+            // attempt to find custom delete action
+            Node& node = selected.nodes.front()->node();
+
+            QKeySequence shortcut = gtApp->getShortCutSequence("delete");
+
+            GtObjectUIAction action = gui_utils::findUIActionByShortCut(node, shortcut);
+            if (action.method())
+            {
+                return action.method()(nullptr, &node);
+            }
+            // node not deletable
+        }
+
+        // create popup to notify that not all nodes are deletable
+        if (QGraphicsView const* view = Impl::activeView(this))
+        {
+            auto* popup = PopupItem::addPopupItem(
+                *this,
+                tr("Selected %1 not deletable!")
+                    .arg(selected.nodes.size() == 1 ? tr("node is") :
+                                                      tr("nodes are")),
+                std::chrono::seconds{1}
+            );
+            // apply cursor position
+            QPointF pos = view->mapToScene(view->mapFromGlobal(QCursor::pos()));
+            pos.rx() -= popup->boundingRect().width() * 0.5;
+            pos.ry() -= popup->boundingRect().height() * 1.5;
+            popup->setPos(pos);
+        }
+    }
+
+    // remove not deletable nodes
+    selected.nodes.erase(beginErase, selected.nodes.end());
+
+    if (selected.empty()) return;
+
+    // default delete selected nodes
     GtObjectList objects;
-    std::transform(selected.connections.begin(), selected.connections.end(), std::back_inserter(objects), [this](ConnectionGraphicsObject* o){ return m_graph->findConnection(o->connectionId()); });
-    std::transform(selected.nodes.begin(), selected.nodes.end(), std::back_inserter(objects), [](NodeGraphicsObject* o){ return &o->node(); });
+    std::transform(selected.connections.begin(),
+                   selected.connections.end(),
+                   std::back_inserter(objects),
+                   [this](ConnectionGraphicsObject* o){
+        return m_graph->findConnection(o->connectionId());
+    });
+    std::transform(selected.nodes.begin(),
+                   selected.nodes.end(),
+                   std::back_inserter(objects),
+                   [](NodeGraphicsObject* o){
+        return &o->node();
+    });
 
     auto modifyCmd = m_graph->modify();
     Q_UNUSED(modifyCmd);
@@ -1015,11 +1110,9 @@ GraphScene::onNodeContextMenu(NodeGraphicsObject* object, QPointF pos)
     // selection should not be empty
     assert(!selected.nodes.empty());
 
-    bool allDeletable = std::all_of(selected.nodes.begin(),
-                                    selected.nodes.end(),
-                                    [](NodeGraphicsObject* o){
-        return o->node().objectFlags() & GtObject::ObjectFlag::UserDeletable;
-    });
+    bool allDeletable = std::none_of(selected.nodes.begin(),
+                                     selected.nodes.end(),
+                                     Impl::isNotDeletable);
 
     Node* selectedNode = &selected.nodes.at(0)->node();
     assert(selectedNode);