+++ src/intelli/nodedatafactory.cpp
@@ -9,6 +9,7 @@
 
 #include "intelli/nodedatafactory.h"
 #include "intelli/nodedata.h"
+#include "intelli/data/invalid.h"
 
 #include "gt_qtutilities.h"
 #include "gt_logging.h"
@@ -47,7 +48,7 @@ struct NodeDataFactory::Impl
 NodeDataFactory::NodeDataFactory() :
     pimpl(std::make_unique<Impl>())
 {
-
+    registerData(GT_METADATA(InvalidData));
 }
 
 NodeDataFactory::~NodeDataFactory() = default;
@@ -100,6 +101,15 @@ NodeDataFactory::registerData(const QMetaObject& meta) noexcept
     }
 
     pimpl->typeNames.insert(className, typeName);
+
+    // register conversion for invalid data type
+    registerConversion(className, typeId<InvalidData>(), [](NodeDataPtr const&){
+        return nullptr;
+    });
+    registerConversion(typeId<InvalidData>(), className, [](NodeDataPtr const&){
+        return nullptr;
+    });
+
     return true;
 }
 
@@ -115,6 +125,14 @@ NodeDataFactory::registerConversion(TypeId const& from,
 
     pimpl->conversions.insert(from, {to, conversion});
     return true;
+}
+
+TypeIdList
+NodeDataFactory::validTypeIds() const
+{
+    TypeIdList list = registeredTypeIds();
+    list.removeOne(typeId<InvalidData>());
+    return list;
 }
 
 TypeName const&