+++ src/intelli/private/graph_impl.h
@@ -11,9 +11,9 @@
 #define GT_INTELLI_GRAPH_IMPL_H
 
 #include <intelli/graph.h>
+#include <intelli/connection.h>
 #include <intelli/nodedatafactory.h>
 #include <intelli/private/utils.h>
-#include <intelli/connection.h>
 
 #include <gt_logging.h>
 
@@ -96,8 +96,8 @@ struct Graph::Impl
                sourceNode->node->parent() == &graph);
 
         // check if ports to connect exist
-        auto inPort  = targetNode->node->port(conId.inPort);
-        auto outPort = sourceNode->node->port(conId.outPort);
+        PortInfo* inPort  = targetNode->node->port(conId.inPort);
+        PortInfo* outPort = sourceNode->node->port(conId.outPort);
 
         if (!inPort || !outPort)
         {
@@ -124,7 +124,7 @@ struct Graph::Impl
         }
 
         // target node should be an input port
-        assert (targetNode->node->portType(inPort->id()) == PortType::In);
+        assert(targetNode->node->portType(inPort->id()) == PortType::In);
 
         // check if types are compatible
         auto& factory = NodeDataFactory::instance();