+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added dummy nodes that are created for missing node classes (dummy objects). Dummy nodes allow to view graphs with unknown nodes and allow the deletion of these nodes. - #107
+
+### Fixed
+- The overlay buttons in a graph scene no longer create undo/redo commands. - #271
+- Nodes that have the `UserDeletable` flag set, can no longer be deleted. A popup is displayed to notify the user in such a case. Custom delete actions are used, identified via the shortcut of an UI action - #270
+- Connections to unknown nodes are now kept and not deleted upon loading a project with missing modules/unknown nodes. - #107
+
 ## [0.13.0] - 2025-03-19
 *This release is not ABI compatible with `0.12.0` and may break API*
 
 ### Added
-- Subgraphs can now also be expanded, thus allowing to "ungroug" subgraphs previously created. - #64
+- Subgraphs can now also be expanded, thus allowing to "ungroup" subgraphs previously created. - #64
 - Allowed nodes to be marked as deprecated. Deprecated nodes should be replaced as soon as possible as a future release may remove the node entirely. - #225
 - Nodes can now have a "display icon" that is drawn in the node header. - #236